fixing siddhi tests
git-svn-id: https://svn.wso2.org/repos/wso2/carbon/platform/trunk@133646 a5903396-d722-0410-b921-86c7d4935375

+++ components/cep/org.wso2.carbon.cep.siddhi/src/test/java/org/wso2/carbon/cep/siddhi/DummyCEPListener.java
@@ -40,7 +40,7 @@ public class DummyCEPListener extends CEPEventListener {
             switch (testNumber) {
                 case 0:
                     mapEvent = (HashMap) event;
-                    Assert.assertTrue(((Integer) mapEvent.get("totalRequestCount") == 21));
+                    Assert.assertTrue(((Long) mapEvent.get("totalRequestCount") == 21));
                     break;
                 case 1:
                     mapEvent = (HashMap) event;
